move files for prismaMock into own directory
