+++ src/venues/venues.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { VenuesService } from './venues.service';
 import { PrismaService } from 'nestjs-prisma';
-import { prismaMock } from 'src/singleton';
+import { prismaMock } from 'src/prismaMock/singleton';
 
 import { venueEntities } from './mocks/venues.mock';
 const mockVenue = venueEntities[0];
